Fixed bug with rendering textures. To set a GLint attribute, you need to use glVertexAttribIPointer, and I was using it like a float. So I just switched the has_texture to float and check if different than 0.0 to see if it has a texture or color

+++ src/renderer.h
@@ -32,7 +32,7 @@ typedef struct Renderer_Vertex {
   Vec4f32 color;
   Vec2f32 uv;
   f32 texture_index;
-  s32 has_texture;
+  f32 has_texture;
 } Renderer_Vertex;
 
 typedef struct Renderer {
